Renamed and moved existing scaffolding engine gh-556
